divided mapper for two classes, sent id from rw to fragment
