+++ app/src/main/java/com/intive/selftraining/selftraining/listmovies/adapter/ItemsAdapter.kt
@@ -9,12 +9,15 @@ import androidx.navigation.Navigation
 import com.intive.selftraining.selftraining.R
 import com.intive.selftraining.selftraining.databinding.ItemViewBinding
 import com.intive.selftraining.selftraining.listmovies.model.ListMoviesMapper
-import com.intive.selftraining.selftraining.movie_details.ItemListener
+import com.intive.selftraining.selftraining.listmovies.model.Results
+import com.intive.selftraining.selftraining.movieDetails.ItemListener
 import com.intive.selftraining.selftraining.network.models.ApiConfiguration
+import android.os.Bundle
+import com.intive.selftraining.selftraining.utils.AppConstants.Companion.ID
 
 class ItemsAdapter : RecyclerView.Adapter<ItemsAdapter.ViewHolder>() {
 
-    private var resultsList: List<ListMoviesMapper.Results> = emptyList()
+    private var resultsList: List<Results> = emptyList()
     private lateinit var apiConfiguration: ApiConfiguration
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
@@ -28,7 +31,7 @@ class ItemsAdapter : RecyclerView.Adapter<ItemsAdapter.ViewHolder>() {
     override fun onBindViewHolder(holder: ViewHolder, position: Int) {
         if (holder is ItemViewHolder && resultsList.size > position) {
             resultsList.let { }
-                holder.bind(resultsList[position])
+            holder.bind(resultsList[position])
         }
     }
 
@@ -48,16 +51,20 @@ class ItemsAdapter : RecyclerView.Adapter<ItemsAdapter.ViewHolder>() {
             false
         )
     ) : ViewHolder(binding.root), ItemListener {
-        override fun onClick(v: View) {
-            Navigation.findNavController(v).navigate(R.id.movieDetailsFragment)
+        var results: Results? = null
+        override fun onClick(view: View) {
+            val args = Bundle()
+            results?.id?.let { args.putInt(ID, it) }
+            Navigation.findNavController(view).navigate(R.id.movieDetailsFragment, args)
         }
 
         fun bind(
-            mapper: ListMoviesMapper.Results
+            results: Results
         ) {
+            this.results = results
             binding.adapter = this
-            binding.text = mapper.title
-            binding.image = mapper.completeImageUrl
+            binding.text = results.title
+            binding.image = results.completeImageUrl
         }
     }
 }
