+++ app/src/main/java/com/intive/selftraining/selftraining/movieDetails/MovieDetailsFragment.kt
@@ -1,4 +1,4 @@
-package com.intive.selftraining.selftraining.movie_details
+package com.intive.selftraining.selftraining.movieDetails
 
 import android.os.Bundle
 import android.support.v4.app.Fragment
@@ -6,20 +6,22 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.intive.selftraining.selftraining.R
+import com.intive.selftraining.selftraining.utils.AppConstants.Companion.ID
+import kotlinx.android.synthetic.main.movies_details_fragment.view.grid_title
 
 class MovieDetailsFragment : Fragment() {
-
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        return inflater.inflate(R.layout.activity_details, container, false)
+        val view = inflater.inflate(R.layout.movies_details_fragment, container, false)
+        view.grid_title.text = arguments?.getInt(ID).toString()
+        return view
     }
 
     override fun onActivityCreated(savedInstanceState: Bundle?) {
         super.onActivityCreated(savedInstanceState)
-
     }
 }
