+++ app/src/main/java/com/intive/selftraining/selftraining/utils/AppConstants.kt
@@ -3,5 +3,6 @@ package com.intive.selftraining.selftraining.utils
 class AppConstants {
     companion object {
         const val URL = "https://api.themoviedb.org/3/"
+        const val ID = "ID"
     }
 }
