+++ app/src/main/res/drawable/photo_background_protection.xml
@@ -1,0 +1,10 @@
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+       android:shape="rectangle">
+    <gradient
+            android:angle="90"
+            android:centerY="0.3"
+            android:startColor="#CC000000"
+            android:centerColor="#66000000"
+            android:endColor="#00000000"
+            android:type="linear"/>
+</shape>
