+++ app/src/main/res/layout/movies_details_fragment.xml
@@ -1,0 +1,130 @@
+<?xml version="1.0" encoding="utf-8"?>
+    <android.support.design.widget.CoordinatorLayout
+            xmlns:android="http://schemas.android.com/apk/res/android"
+            xmlns:app="http://schemas.android.com/apk/res-auto"
+            xmlns:tools="http://schemas.android.com/tools"
+            android:fitsSystemWindows="true"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent">
+        <android.support.design.widget.AppBarLayout
+                android:id="@+id/main.appbar"
+                android:fitsSystemWindows="true"
+                android:layout_width="match_parent"
+                android:layout_height="@dimen/app_bar_height"
+                android:theme="@style/ThemeOverlay.AppCompat.Dark.ActionBar">
+
+            <android.support.design.widget.CollapsingToolbarLayout
+                    android:id="@+id/main_collapsing"
+                    android:layout_width="match_parent"
+                    android:layout_height="match_parent"
+                    app:contentScrim="@color/colorPrimary"
+                    app:layout_scrollFlags="scroll|exitUntilCollapsed">
+
+                <ImageView
+                        android:id="@+id/main_backdrop"
+                        android:layout_width="match_parent"
+                        android:layout_height="match_parent"
+                        android:fitsSystemWindows="true"
+                        android:scaleType="centerCrop"
+                        android:transitionName="poster"
+                        app:layout_collapseMode="parallax"/>
+
+                <View
+                        android:layout_width="match_parent"
+                        android:layout_height="@dimen/text_protection_scrim_height"
+                        android:layout_alignBottom="@+id/main_backdrop"
+                        android:layout_gravity="bottom"
+                        android:background="@drawable/photo_background_protection"/>
+
+                <android.support.v7.widget.Toolbar
+                        android:id="@+id/main_toolbar"
+                        android:layout_width="match_parent"
+                        android:layout_height="?attr/actionBarSize"
+                        app:layout_collapseMode="pin"
+                        app:popupTheme="@style/ThemeOverlay.AppCompat.Light"/>
+            </android.support.design.widget.CollapsingToolbarLayout>
+        </android.support.design.widget.AppBarLayout>
+
+        <android.support.v4.widget.NestedScrollView
+                android:layout_width="match_parent"
+                android:layout_height="match_parent"
+                android:padding="@dimen/activity_horizontal_margin"
+                app:layout_behavior="@string/appbar_scrolling_view_behavior">
+            <android.support.v7.widget.GridLayout
+                    xmlns:android="http://schemas.android.com/apk/res/android"
+                    android:layout_width="match_parent"
+                    app:useDefaultMargins="true"
+                    android:layout_height="match_parent"
+                    app:columnCount="2">
+                <TextView
+                        app:layout_column="0"
+                        app:layout_row="0"
+                        android:text="@string/title"
+                        style="@style/left_column_bold"/>
+
+                <TextView
+                        android:id="@+id/grid_title"
+                        style="@style/right_column_normal"
+                        app:layout_column="1"
+                        app:layout_row="0"
+                        tools:text="@string/example_title"/>
+                <TextView
+                        style="@style/left_column_bold"
+                        app:layout_column="0"
+                        app:layout_row="1"
+                        android:text="@string/released"/>
+                <TextView
+                        style="@style/right_column_normal"
+                        app:layout_column="1"
+                        app:layout_row="1"
+                        tools:text="@string/example_release"
+                        android:id="@+id/grid_released"/>
+
+                <TextView
+                        style="@style/left_column_bold"
+                        android:text="@string/vote_average"
+                        app:layout_column="0"
+                        app:layout_row="2"/>
+
+                <TextView
+                        android:id="@+id/grid_votes"
+                        style="@style/right_column_normal"
+                        app:layout_column="1"
+                        app:layout_row="2"/>
+                <TextView
+                        style="@style/left_column_bold"
+                        app:layout_column="0"
+                        app:layout_row="4"
+                        android:text="@string/Genre"/>
+                <TextView
+                        style="@style/right_column_normal"
+                        app:layout_column="1"
+                        app:layout_row="4"
+                        tools:text="@string/example_genre"
+                        android:id="@+id/grid_genre"/>
+                <TextView
+                        app:layout_column="0"
+                        app:layout_row="7"
+                        style="@style/left_column_bold"
+                        android:text="@string/plot"/>
+                <TextView
+                        app:layout_column="1"
+                        app:layout_row="7"
+                        tools:text="@string/example_full_plot"
+                        style="@style/right_column_normal"
+                        android:id="@+id/grid_plot"/>
+            </android.support.v7.widget.GridLayout>
+
+        </android.support.v4.widget.NestedScrollView>
+
+        <android.support.design.widget.FloatingActionButton
+                android:id="@+id/fab"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_margin="16dp"
+                android:clickable="true"
+                android:src="@drawable/movie"
+                app:backgroundTint="@color/colorPrimaryDark"
+                app:layout_anchor="@id/main.appbar"
+                app:layout_anchorGravity="bottom|right|end"/>
+    </android.support.design.widget.CoordinatorLayout>
