+++ app/src/main/res/navigation/main_graph.xml
@@ -8,6 +8,6 @@
                 app:destination="@id/movieDetailsFragment"/>
     </fragment>
     <fragment android:id="@+id/movieDetailsFragment"
-              android:name="com.intive.selftraining.selftraining.movie_details.MovieDetailsFragment"
+              android:name="com.intive.selftraining.selftraining.movieDetails.MovieDetailsFragment"
               android:label="MovieDetailsFragment"/>
 </navigation>
